Aggregating heartbeats functionality into a feature
