+++ src/ServiceControl/ServiceControl.csproj
@@ -343,8 +343,7 @@
     <Compile Include="EventLog\IEventLogMappingDefinition.cs" />
     <Compile Include="EventLog\EventLogItem.cs" />
     <Compile Include="HeartbeatMonitoring\HeartbeatingEndpoint.cs" />
-    <Compile Include="HeartbeatMonitoring\HeartbeatMonitor.cs" />
-    <Compile Include="HeartbeatMonitoring\HeartbeatStatusInitializer.cs" />
+    <Compile Include="HeartbeatMonitoring\HeartbeatsMonitor.cs" />
     <Compile Include="HeartbeatMonitoring\HeartbeatsIndex.cs" />
     <Compile Include="HeartbeatMonitoring\HeartbeatStatus.cs" />
     <Compile Include="HeartbeatMonitoring\HeartbeatStatusProvider.cs" />
